+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/persistence/TemplateDAO.java
@@ -87,7 +87,7 @@ public class TemplateDAO extends BaseDAO {
         return retrieveObjects(query);
     }
 
-    public Long count(String query) throws DAOException {
+    public Integer count(String query) throws DAOException {
         return retrieveAmount(query);
     }
 }
