+++ Kitodo/src/main/java/org/kitodo/services/data/TaskService.java
@@ -23,7 +23,6 @@ import java.util.List;
 import java.util.Set;
 
 import org.hibernate.Session;
-import org.kitodo.data.database.beans.Property;
 import org.kitodo.data.database.beans.Task;
 import org.kitodo.data.database.exceptions.DAOException;
 import org.kitodo.data.database.helper.enums.TaskStatus;
@@ -112,7 +111,7 @@ public class TaskService extends TitleSearchService<Task> {
         return taskDAO.search(query);
     }
 
-    public Long count(String query) throws DAOException {
+    public Integer count(String query) throws DAOException {
         return taskDAO.count(query);
     }
